rdb file format implementation

+++ src/main/tonivade/db/command/zset/SortedSetAddCommand.java
@@ -35,7 +35,7 @@ public class SortedSetAddCommand implements ICommand {
             DatabaseValue initial = db.getOrDefault(request.getParam(0), zset());
             DatabaseValue result = db.merge(request.getParam(0), parseInput(request),
                     (oldValue, newValue) -> {
-                        Set<Entry<Float, String>> merge = new SortedSet();
+                        Set<Entry<Double, String>> merge = new SortedSet();
                         merge.addAll(oldValue.getValue());
                         merge.addAll(newValue.getValue());
                         return zset(merge);
@@ -51,7 +51,7 @@ public class SortedSetAddCommand implements ICommand {
     }
 
     private DatabaseValue parseInput(IRequest request) throws NumberFormatException {
-        Set<Entry<Float, String>> set = new SortedSet();
+        Set<Entry<Double, String>> set = new SortedSet();
         String score = null;
         for (String string : request.getParams().stream().skip(1).collect(Collectors.toList())) {
             if (score != null) {
